+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix StreakPanel calendar UI [#3910](https://github.com/rokwire/illinois-app/issues/3910).
 
 ## [5.1.17] - 2024-01-30
 ### Added