+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Fix StreakPanel calendar UI [#3910](https://github.com/rokwire/illinois-app/issues/3910).
+- Show "Group Members Only" label for group events [#3912](https://github.com/rokwire/illinois-app/issues/3912).
 
 ## [5.1.17] - 2024-01-30
 ### Added