+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated display strings when linking event to a group, polished GroupsSelectionPopup UI [#3905](https://github.com/rokwire/illinois-app/issues/3905).
 ### Added
 - Show published status in event detail panel for admins only [#3908](https://github.com/rokwire/illinois-app/issues/3908).
+- Added published info link in Events filters panel [#3886](https://github.com/rokwire/illinois-app/issues/3886).
 
 ## [5.1.17] - 2024-01-30
 ### Added