+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Essential Skills Coach Dashboard [#3425](https://github.com/rokwire/illinois-app/issues/3425).
+
+## [5.1.18] - 2024-01-31
 ### Fixed
 - Fix StreakPanel calendar UI [#3910](https://github.com/rokwire/illinois-app/issues/3910).
 - Show "Group Members Only" label for group events [#3912](https://github.com/rokwire/illinois-app/issues/3912).